+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - **Configuration**
   - Added `[tools.*.plugin]` and `[backends.*.plugin]` fields for specifying a plugin locator for the respective tool or backend. This is an alternative to the `[plugins.tools]` and `[plugins.backends]` tables.
 - **Lockfiles**
+  - Added a "Locked" column to the `proto outdated` and `proto status` tables, which displays the version locked for the configured spec. The column is only displayed when a tool has a matching lockfile record. Additionally, both commands now include a `locked_version` field in their `--json` output.
   - Updated `proto outdated --update` to also update matching records in the lockfile when versions are written to their respective configs. Since the new versions have not been installed yet, checksums are removed from the migrated records, and will be re-populated on the next install.
 - **WASM API**
   - Added a `RealPath` type, which is a newtype wrapper around `PathBuf` that represents a real path on the host file system. This is a sibling to the `VirtualPath` type, which represents a virtual path in the guest WASM environment.